+++ src/libromdata/Audio/BCSTM.hpp
@@ -1,0 +1,34 @@
+/***************************************************************************
+ * ROM Properties Page shell extension. (libromdata)                       *
+ * BCSTM.hpp: Nintendo 3DS BCSTM audio reader.                             *
+ *                                                                         *
+ * Copyright (c) 2019 by David Korth.                                      *
+ *                                                                         *
+ * This program is free software; you can redistribute it and/or modify it *
+ * under the terms of the GNU General Public License as published by the   *
+ * Free Software Foundation; either version 2 of the License, or (at your  *
+ * option) any later version.                                              *
+ *                                                                         *
+ * This program is distributed in the hope that it will be useful, but     *
+ * WITHOUT ANY WARRANTY; without even the implied warranty of              *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           *
+ * GNU General Public License for more details.                            *
+ *                                                                         *
+ * You should have received a copy of the GNU General Public License       *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.   *
+ ***************************************************************************/
+
+#ifndef __ROMPROPERTIES_LIBROMDATA_AUDIO_BCSTM_HPP__
+#define __ROMPROPERTIES_LIBROMDATA_AUDIO_BCSTM_HPP__
+
+#include "librpbase/RomData.hpp"
+
+namespace LibRomData {
+
+ROMDATA_DECL_BEGIN(BCSTM)
+ROMDATA_DECL_METADATA()
+ROMDATA_DECL_END()
+
+}
+
+#endif /* __ROMPROPERTIES_LIBROMDATA_AUDIO_BCSTM_HPP__ */