+++ NEWS.md
@@ -42,6 +42,7 @@
     XEX executables is supported at the moment.
   * Xbox360_XEX: Xbox 360 executables.
   * BRSTM: Nintendo Wii BRSTM audio files.
+  * BCSTM: Nintendo 3DS BCSTM audio files.
 
 * New audio parsers:
   * ADX: CRI ADX audio format. Used by many Sega games.